Extract prepareParameterKeyMapping to trait
