+++ src/Mysqli/MysqliStatement.php
@@ -12,6 +12,7 @@ use Joomla\Database\Exception\ExecutionFailureException;
 use Joomla\Database\Exception\PrepareStatementFailureException;
 use Joomla\Database\FetchMode;
 use Joomla\Database\FetchOrientation;
+use Joomla\Database\ParameterKeyMappingTrait;
 use Joomla\Database\ParameterType;
 use Joomla\Database\StatementInterface;
 
@@ -24,6 +25,8 @@ use Joomla\Database\StatementInterface;
  */
 class MysqliStatement implements StatementInterface
 {
+	use ParameterKeyMappingTrait;
+
 	/**
 	 * Values which have been bound to the statement.
 	 *
@@ -125,137 +128,6 @@ class MysqliStatement implements StatementInterface
 		}
 
 		$this->statement = $statement;
-	}
-
-	/**
-	 * Replace named parameters with numbered parameters
-	 *
-	 * @param   string  $sql  The SQL statement to prepare.
-	 *
-	 * @return  string  The processed SQL statement.
-	 *
-	 * @since   2.0.0
-	 */
-	private function prepareParameterKeyMapping(string $sql): string
-	{
-		$startPos  	= 0;
-		$literal    = '';
-		$mapping    = [];
-		$matches    = [];
-		$pattern    = '/([:][a-zA-Z0-9_]+)/';
-
-		if (!preg_match($pattern, $sql, $matches))
-		{
-			return $sql;
-		}
-
-		$sql = trim($sql);
-		$n   = \strlen($sql);
-
-		while ($startPos < $n)
-		{
-			if (!preg_match($pattern, $sql, $matches, 0, $startPos))
-			{
-				break;
-			}
-
-			$j = strpos($sql, "'", $startPos);
-			$k = strpos($sql, '"', $startPos);
-
-			if (($k !== false) && (($k < $j) || ($j === false)))
-			{
-				$quoteChar = '"';
-				$j         = $k;
-			}
-			else
-			{
-				$quoteChar = "'";
-			}
-
-			if ($j === false)
-			{
-				$j = $n;
-			}
-
-			// Search for named prepared parameters and replace it with ? and save its position
-			$substring = substr($sql, $startPos, $j - $startPos);
-
-			if (preg_match_all($pattern, $substring, $matches, PREG_PATTERN_ORDER + PREG_OFFSET_CAPTURE))
-			{
-				foreach ($matches[0] as $i => $match)
-				{
-					if ($i === 0)
-					{
-						$literal .= substr($substring, 0, $match[1]);
-					}
-
-					$mapping[$match[0]]     = \count($mapping);
-					$endOfPlaceholder       = $match[1] + strlen($match[0]);
-					$beginOfNextPlaceholder = $matches[0][$i + 1][1] ?? strlen($substring);
-					$beginOfNextPlaceholder -= $endOfPlaceholder;
-					$literal                .= '?' . substr($substring, $endOfPlaceholder, $beginOfNextPlaceholder);
-				}
-			}
-			else
-			{
-				$literal .= $substring;
-			}
-
-			$startPos = $j;
-			$j++;
-
-			if ($j >= $n)
-			{
-				break;
-			}
-
-			// Quote comes first, find end of quote
-			while (true)
-			{
-				$k       = strpos($sql, $quoteChar, $j);
-				$escaped = false;
-
-				if ($k === false)
-				{
-					break;
-				}
-
-				$l = $k - 1;
-
-				while ($l >= 0 && $sql[$l] === '\\')
-				{
-					$l--;
-					$escaped = !$escaped;
-				}
-
-				if ($escaped)
-				{
-					$j = $k + 1;
-
-					continue;
-				}
-
-				break;
-			}
-
-			if ($k === false)
-			{
-				// Error in the query - no end quote; ignore it
-				break;
-			}
-
-			$literal .= substr($sql, $startPos, $k - $startPos + 1);
-			$startPos = $k + 1;
-		}
-
-		if ($startPos < $n)
-		{
-			$literal .= substr($sql, $startPos, $n - $startPos);
-		}
-
-		$this->parameterKeyMapping = $mapping;
-
-		return $literal;
 	}
 
 	/**
